fix(FR-3514): render the modal loading body as antd's paragraph skeleton
Both modal implementations painted their `loading` body with the Astryx
`Skeleton` primitive at a fixed `height={120}` — one flat grey slab, because
Astryx's `Skeleton` is a single box and antd's multi-part shimmer had to be
composed by hand at migration time. `BAISkeletonAstryx` is that composition and
already backs 74 app-side loading fallbacks, but it lived in `react/src`, which
BUI cannot import from, so `BAIModal` had no way to reach it.

Move `BAISkeletonAstryx` into `packages/backend.ai-ui`, export it from the BUI
barrel, and leave the old app-side path as a re-export so the 74 call sites are
untouched. Both `BAIModal` (BUI) and `BAIModalAstryx` (react) now render
`hasTitle={false} rows={4}`, which is what antd's `Modal` rendered — the title
bar is off because the dialog header already carries the title.
